chore: Update consumer index.html to use unpkg CDN for picard-js
This commit updates the consumer index.html file to use the unpkg CDN for the picard-js library instead of the jsDelivr CDN. The jsDelivr CDN script tag has been commented out, and a new script tag pointing to the unpkg CDN has been added. This change was made to improve the loading performance and ensure the availability of the picard-js library.